[#14] Separate out failover tests
These are the least likely to be re-used; either by our other services or by
third-parties using our open source code. So don't clutter the main test
file with them. Since they all have consistent naming they can be selected
with `-run TestFailover`.
