+++ cdn_acceptance_test.go
@@ -170,48 +170,6 @@ func TestDefaultTTL(t *testing.T) {
 	t.Error("Not implemented")
 }
 
-// Should serve stale object and not hit mirror(s) if origin is down and
-// object is beyond TTL but still in cache.
-func TestFailoverOriginDownServeStale(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should serve stale object and not hit mirror(s) if origin returns a 5xx
-// response and object is beyond TTL but still in cache.
-func TestFailoverOrigin5xxServeStale(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should fallback to first mirror if origin is down and object is not in
-// cache (active or stale).
-func TestFailoverOriginDownUseFirstMirror(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should fallback to first mirror if origin returns 5xx response and object
-// is not in cache (active or stale).
-func TestFailoverOrigin5xxUseFirstMirror(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should fallback to second mirror if both origin and first mirror are
-// down.
-func TestFailoverOriginDownFirstMirrorDownUseSecondMirror(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should fallback to second mirror if both origin and first mirror return
-// 5xx responses.
-func TestFailoverOrigin5xxFirstMirror5xxUseSecondMirror(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should not fallback to mirror if origin returns a 5xx response with a
-// No-Fallback header.
-func TestFailoverNoFallbackHeader(t *testing.T) {
-	t.Error("Not implemented")
-}
-
 // Should serve a known static error page if cannot serve a page
 // from origin, stale or any mirror.
 // NB: ideally this should be a page that we control that has a mechanism
